moved appprotect validation into validation package
